remove int helpers
