+++ deployment/environment.go
@@ -6,7 +6,6 @@ import (
 	"encoding/hex"
 	"errors"
 	"fmt"
-	"math/big"
 	"regexp"
 	"slices"
 	"sort"
@@ -26,22 +25,6 @@ import (
 
 	"github.com/smartcontractkit/chainlink-common/keystore/corekeys/p2pkey"
 )
-
-func UBigInt(i uint64) *big.Int {
-	return new(big.Int).SetUint64(i)
-}
-
-func E18Mult(amount uint64) *big.Int {
-	return new(big.Int).Mul(UBigInt(amount), UBigInt(1e18))
-}
-
-// EDecMult scales amount by the number of decimals
-func EDecMult(amount uint64, decimals int64) *big.Int {
-	return new(big.Int).Mul(
-		UBigInt(amount),
-		new(big.Int).Exp(big.NewInt(10), big.NewInt(decimals), nil),
-	)
-}
 
 type OCRConfig struct {
 	OffchainPublicKey types2.OffchainPublicKey
