+++ integration-tests/smoke/ccip/ccip_disable_lane_test.go
@@ -13,12 +13,12 @@ import (
 	"github.com/smartcontractkit/chainlink-ccip/chains/evm/gobindings/generated/v1_6_0/onramp"
 	cldf_chain "github.com/smartcontractkit/chainlink-deployments-framework/chain"
 	"github.com/smartcontractkit/chainlink-testing-framework/lib/utils/testcontext"
-	"github.com/smartcontractkit/chainlink/deployment"
 	"github.com/smartcontractkit/chainlink/deployment/ccip/changeset/testhelpers"
 	"github.com/smartcontractkit/chainlink/deployment/ccip/changeset/v1_6"
 	ccipclient "github.com/smartcontractkit/chainlink/deployment/ccip/shared/client"
 	"github.com/smartcontractkit/chainlink/deployment/ccip/shared/stateview"
 	testsetups "github.com/smartcontractkit/chainlink/integration-tests/testsetups/ccip"
+	"github.com/smartcontractkit/chainlink/integration-tests/utils/bigint"
 )
 
 // Intention of this test is to ensure that the lane can be disabled and enabled correctly
@@ -43,8 +43,8 @@ func TestDisableLane(t *testing.T) {
 		expectedSeqNumExec     = make(map[testhelpers.SourceDestPair][]uint64)
 		startBlocks            = make(map[uint64]*uint64)
 		pairs                  []testhelpers.SourceDestPair
-		linkPrice              = deployment.E18Mult(100)
-		wethPrice              = deployment.E18Mult(4000)
+		linkPrice              = bigint.E18Mult(100)
+		wethPrice              = bigint.E18Mult(4000)
 		noOfRequests           = 3
 		sendmessage            = func(src, dest uint64, deployer *bind.TransactOpts) (*onramp.OnRampCCIPMessageSent, error) {
 			out, err := testhelpers.SendRequest(
